WIP: partially working authentication, need to finalize account log-out
