+++ DungeonsandDragons/app/src/main/AndroidManifest.xml
@@ -25,8 +25,7 @@
                 <data android:mimeType="text/plain"/>
             </intent-filter>
         </activity>
-
-        <activity android:name=".Signup"/>
+        <activity android:name=".Startup"/>
 <!--        <activity android:name=".NFC">-->
 <!--            <intent-filter>-->
 <!--                <action android:name="android.nfc.action.NDEF_DISCOVERED"/>-->
