+++ README.md
@@ -1,1 +1,7 @@
 # Dungeons and Dragons web application
+
+## Notes
+* Need to start emulator for firebase for local development, use
+```bat
+firebase emulators:start
+```
